+++ src/main/java/modelo/dao/ProductoDAO.java
@@ -7,7 +7,7 @@ public interface ProductoDAO {
     public void create(Producto p) throws Exception;
     public List<Producto> read() throws Exception;
     public void update(Producto p) throws Exception;
-    public void delete(int id) throws Exception;
+    public void delete(long id) throws Exception;
     
     public Producto obtenerPorNombre(String nombre) throws Exception;
     public List<String> obtenerNombresProductosOrdenados() throws Exception;
